% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_functions.R
\name{apply_rating}
\alias{apply_rating}
\title{Apply ratings and points to a dataframe}
\usage{
apply_rating(
  df,
  score_col,
  indicator_col,
  measure_col,
  emh_col,
  subject_col,
  n_col,
  category_col,
  cuts_df,
  rating_col = "rating",
  pts_col = "pts"
)
}
\arguments{
\item{df}{A dataframe with score, indicator, measure, EMH, subject, N-count, and category columns}

\item{score_col}{Name of the column with scores}

\item{indicator_col}{Name of the column with indicator}

\item{measure_col}{Name of the column with measure}

\item{emh_col}{Name of the column with EMH codes}

\item{subject_col}{Name of the column with subject}

\item{n_col}{Name of the column with N-count}

\item{category_col}{Name of the column with category}

\item{cuts_df}{Cut scores dataframe}

\item{rating_col}{Name of the column to create for ratings (default "rating")}

\item{pts_col}{Name of the column to create for points (default "pts")}
}
\value{
Dataframe with added rating and pts columns
}
\description{
Apply ratings and points to a dataframe
}
