% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cde-loss.R
\name{cde_loss}
\alias{cde_loss}
\title{Calculate conditional density estimation loss.}
\usage{
cde_loss(cdes, z_grid, z_test)
}
\arguments{
\item{cdes}{a matrix of conditional density estimates with each row
corresponding to an observation, and each column corresponding to
a grid point}

\item{z_grid}{a matrix of the grid points at which the conditional
densities are estimated}

\item{z_test}{a matrix of the true responses}
}
\description{
Estimates the CDE loss
\deqn{\int \int (\hat{f}(z \mid x) - f(z \mid x))^{2} dz dP(x)}
which can be written as
\deqn{E[\int \hat{f}(z \mid X) dz] - 2 E[\hat{f}(Z \mid X)] + C_{f}}
where \eqn{C_{f}} is an (unknown) constant. These expectations are
estimated using empirical expectations.
}
\examples{
z_grid <- seq(0, 1, length.out = 99)
z_test <- runif(101)
cde <- matrix(dbeta(z_grid, 2, 2), nrow = 101, ncol = 99, byrow=TRUE)
cde_loss(cde, z_grid, z_test)
}
