% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_coverage.R
\name{cdf_coverage}
\alias{cdf_coverage}
\title{Calculates coverage based upon the CDF}
\usage{
cdf_coverage(cdes, z_grid, z_test)
}
\arguments{
\item{cdes}{A matrix of conditional density estimates; each row
corresponds to an observation, each column corresponds to a grid
point}

\item{z_grid}{A vector/matrix of grid points at which the CDE is
estimated}

\item{z_test}{The true responses associated with the observations}
}
\value{
A vector of "p-values"
}
\description{
Calculates coverage based upon the CDF
}
\examples{
  set.seed(12345)
  n_grid <- 10001
  n_test <- 73
  z_grid <- seq(-5, 5, length.out = n_grid)
  z_test <- rnorm(n_test)
  cdes <- matrix(dnorm(z_grid), nrow = n_test, ncol = n_grid, byrow = TRUE)
  pvals <- cdf_coverage(cdes, z_grid, z_test)
  ks.test(pvals, "punif")
}
