% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_uniq.R
\name{check_uniq}
\alias{check_uniq}
\title{Check uniqueness of a dataset by specified variables}
\usage{
check_uniq(df, vars, n_show = 5, df_label = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{vars}{A character vector of variable names to check uniqueness by.}

\item{n_show}{Number of duplicate rows to show (default is 5).}

\item{df_label}{Name of df being checked.}
}
\value{
Returns the original data frame if unique by \code{vars}; otherwise throws an error.
}
\description{
This function checks whether the input dataset is unique by a given set of variables.
If duplicates are found, the function prints a sample and stops with an error.
}
