% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_functions.R
\name{get_cut_scores}
\alias{get_cut_scores}
\title{Get Cut Scores (Default or External)}
\usage{
get_cut_scores(path = NULL)
}
\arguments{
\item{path}{Optional path to Excel file. If NULL, uses internal package data.}
}
\value{
A list of cut scores
}
\description{
Get Cut Scores (Default or External)
}
