% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_functions.R
\name{get_pts}
\alias{get_pts}
\title{Get points from rating}
\usage{
get_pts(rating, category)
}
\arguments{
\item{rating}{Rating value}

\item{category}{Category type}
}
\value{
Numeric points earned
}
\description{
Get points from rating
}
