% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_functions.R
\name{get_rating}
\alias{get_rating}
\title{Get rating based on score and cut scores table}
\usage{
get_rating(score, indicator, measure, emh_code, subject, cuts_df, n)
}
\arguments{
\item{score}{Numeric score to rate}

\item{indicator}{Indicator (e.g. "GRO", "ACH")}

\item{measure}{Measure (e.g. "GROWTH", "WIDA OTG")}

\item{emh_code}{EMH level ("E", "M", "H")}

\item{subject}{Subject (e.g. "ELA", "MATH")}

\item{cuts_df}{Cut scores dataframe}

\item{n}{N-count or rule for missing data}
}
\value{
Rating ("Exceeds", "Meets", "Approaching", "Does Not Meet", "-")
}
\description{
Get rating based on score and cut scores table
}
