% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd_contours.R
\name{hpd_contour_levels}
\alias{hpd_contour_levels}
\title{Calculate contour levels for HPD sets}
\usage{
hpd_contour_levels(density, z_grid, probs, max_iters = 200,
  tol = 0.001)
}
\arguments{
\item{density}{A vector of density evaluations.}

\item{z_grid}{The grid on which the density is evaluated. Should be
regularly spaced.}

\item{probs}{The probability levels for the HPD sets}

\item{max_iters}{Optional; number of iterations of binary search}

\item{tol}{Optional; tolerance of binary search}
}
\description{
Calculate contour levels for HPD sets
}
