% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd_coverage.R
\name{hpd_coverage}
\alias{hpd_coverage}
\title{Calculate coverage based upon HPD regions.}
\usage{
hpd_coverage(cdes, z_grid, z_test)
}
\arguments{
\item{cdes}{A matrix of conditional density estimates; each row
corresponds to an observation, each column corresponds to a grid
point}

\item{z_grid}{A matrix of grid points at which the CDE is
estimated}

\item{z_test}{The true responses associated with the observations}
}
\value{
A vector of values; under the true generating model
  these are Uniform(0, 1)
}
\description{
Calculate coverage based upon HPD regions.
}
\examples{
  set.seed(12345)
  n_grid <- 10001
  n_test <- 73
  z_grid <- seq(-5, 5, length.out = n_grid)
  z_test <- rnorm(n_test)
  cdes <- matrix(dnorm(z_grid), nrow = n_test, ncol = n_grid, byrow = TRUE)
  vals <- hpd_coverage(cdes, z_grid, z_test)
  ks.test(vals, "punif")
}
