% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sql.R
\name{query_sql}
\alias{query_sql}
\title{Load SQL from file, apply glue substitution, and run query}
\usage{
query_sql(con, filepath, settings = list())
}
\arguments{
\item{con}{A DBI database connection}

\item{filepath}{Path to the SQL file}

\item{settings}{A named list of values for glue substitution}
}
\value{
A data frame with the query results
}
\description{
Load SQL from file, apply glue substitution, and run query
}
