% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_functions.R
\name{read_cut_scores}
\alias{read_cut_scores}
\title{Read Cut Scores from Excel}
\usage{
read_cut_scores(path)
}
\arguments{
\item{path}{Path to Excel file with cut scores.}
}
\value{
A list with traditional and AEC cut scores as tibbles.
}
\description{
Read Cut Scores from Excel
}
