% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_sch_data.R
\docType{data}
\name{samp_sch_data}
\alias{samp_sch_data}
\title{samp_sch_data}
\format{
A data frame with 26 rows and 16 columns:
\describe{
  \item{DASY_KEY}{School year.}
  \item{DETAIL_DASY_KEY}{Detail school year.}
  ...
}
}
\usage{
data(samp_sch_data)
}
\description{
A sample dataset for demonstration purposes. This data is not real.
}
\keyword{datasets}
