% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_functions..R
\name{trans_sum}
\alias{trans_sum}
\title{Summarize rating changes (before→after) for one test/subgroup}
\usage{
trans_sum(df, test = NULL, subgroup = NULL, debug = FALSE)
}
\arguments{
\item{df}{A data frame containing at least \code{test}, \code{sub}, \code{rating_pre}, and \code{rating_post}}

\item{test}{Optional string to filter on \code{test}}

\item{subgroup}{Optional string to filter on \code{sub}}

\item{debug}{Logical; if \code{TRUE}, prints any rows where change is \code{NA}}
}
\value{
An HTML table (a \code{kable}) with counts & percents by change category
}
\description{
Summarize rating changes (before→after) for one test/subgroup
}
