% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_col_map.R
\name{apply_col_map}
\alias{apply_col_map}
\title{Apply Column Mapping and Convert Types}
\usage{
apply_col_map(df)
}
\arguments{
\item{df}{A dataframe whose columns need renaming.}
}
\value{
A dataframe with renamed columns.
}
\description{
This function renames columns in a dataframe based on a predefined mapping
and can also apply type conversions if needed.
}
\examples{
df <- data.frame(DASY_KEY = 2022, DDST_DIST_NUMBER = 1234)
apply_col_map(df)
}
