% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_sch_data.R
\docType{data}
\name{samp_sch_data}
\alias{samp_sch_data}
\title{samp_sch_data}
\format{
A data frame with 26 rows and 16 columns:
\describe{
\item{DASY_KEY}{School year.}
\item{DETAIL_DASY_KEY}{Detail school year.}
\item{DDST_DIST_NUMBER}{District number.}
\item{DDST_DISTRICT_NAME}{District name.}
\item{DSCH_SCHOOL_NAME}{School name.}
\item{DSCH_SCHOOL_NUMBER}{School number.}
\item{INCLUDED_IN_CALCS_YN}{Included in calcs?}
\item{INCLUSION_RULE}{Inclusion rule.}
\item{DATA_SET}{Data set.}
\item{YEAR_SPAN}{Year span.}
\item{SUBJECT}{Subject.}
\item{SASID}{Student ID.}
\item{SCALE_SCORE}{Scale score.}
\item{MEDIAN_SGP}{Median SGP.}
\item{VALID_SCORE}{Valid score.}
\item{DEMH_EMH_CODE}{EMH code.}
}
}
\usage{
data(samp_sch_data)
}
\description{
A sample dataset for demonstration purposes. This data is not real.
}
\keyword{datasets}
